move logic to reusable SpanExtensions
